02_group_B/requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.1
pydantic>=2.6.0
typing-extensions>=4.9.0
tabulate>=0.9.0

